make form data changeable on live preview
